404.php + htaccess

+++ controller/deleteReports.php
@@ -1,7 +1,7 @@
 <?php
 require_once($_SERVER['DOCUMENT_ROOT']."/site/model/implementation/Database.php");
 
-if (isset($_POST['idReported']) && isset($_POST['typeReported'])) {
+if (isset($session->admin) && isset($_POST['idReported']) && isset($_POST['typeReported'])) {
     $bdd = Database::getConnection();
     $bdd->exec('DELETE FROM Reporting WHERE idReported = \''.$_POST['idReported'].'\' AND typeReported = \''.$_POST['typeReported'].'\'');  
 }
